pyproject.toml
# package build
# https://peps.python.org/pep-0517/
[build-system]
requires = ["setuptools>=65", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

# package basics
# https://peps.python.org/pep-0621/
[project]
name = "pyeil"
description = "Extract imported libraries from Python, R, and other source code files using tree-sitter."
keywords = []
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MPLv2 License" }
authors = [
  { email = "evamaxfieldbrown@gmail.com", name = "Eva Maxfield Brown" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dynamic = ["version"]
dependencies = [
  "pyyaml>=6",
  "stdlib-list>=0.12",
  "tqdm>=4",
  "tree-sitter>=0.20",
  "tree-sitter-language-pack>=0.13",
]

[project.urls]
Homepage = "https://github.com/evamaxfield/py-extract-imported-libraries"
"Bug Tracker" = "https://github.com/evamaxfield/py-extract-imported-libraries/issues"
Documentation = "https://evamaxfield.github.io/py-extract-imported-libraries"
"User Support" = "https://github.com/evamaxfield/py-extract-imported-libraries/issues"

# extra dependencies
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
dev = [
  "ipython",
  "jupyterlab",
]
lint = [
  "prek>=0.2.29",
  "ty>=0.0.12",
]
test = [
  "pytest>=9",
]

# entry points
# https://peps.python.org/pep-0621/#entry-points
# [project.entry-points."console_scripts"]

# build settings
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["*docs/*", "*tests/*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "py.typed", "data/stdlibs.yml"]

# tools
# https://github.com/charliermarsh/ruff
[tool.ruff]
line-length = 96
target-version = "py312"
lint.extend-select = [
    "E",      # style errors
    "F",      # flakes
    "D",      # pydocstyle
    "I001",   # isort
    "UP",     # pyupgrade
    "N",      # pep8-naming
    # "S",    # bandit
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "A001",   # flake8-builtins
    "RUF",    # ruff-specific rules
    "RUF100", # Unused noqa directive
]
lint.extend-ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D103", # Missing docstring in public function
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D413", # Missing blank line after last section
    "D416", # Section name should end with a colon
]

[tool.ruff.lint.per-file-ignores]
"**/notebooks/*" = ["E501", "B018"]
"**/publications/*" = ["E501", "B018"]
"**/tests/*.py" = ["D"]
"**/migrations/*.py" = ["D"]

[tool.ty.src]
exclude = [".github", ".pre-commit-config.yaml", "pyproject.toml"]